+++ building/genCloudComPy_Conda310_Ubuntu2004.sh
@@ -14,7 +14,6 @@ fi
 export CONDA_ENV=CloudComPy310                                                         # conda environment name
 export CONDA_PATH=${CONDA_ROOT}/envs/${CONDA_ENV}                                      # conda environment directory
 export CORK_REP=${HOME}/projets/CloudComPy/cork                                        # directory of cork (remove the plugin in cmake options if not needed)
-export DRACO_REP=${HOME}/projets/CloudComPy/dracoInstall                               # directory of draco (remove the plugin in cmake options if not needed)
 export FBXSDK_REP=${HOME}/projets/CloudComPy/fbxSdk                                    # directory of fbx sdk (remove the plugin in cmake options if not needed)
 export LIBIGL_REP=${HOME}/projets/CloudComPy/libigl                                    # directory of libigl (remove the plugin in cmake options if not needed)
 if [ -d ${HOME}/projets/CloudComPy/Occ-740p3_opt ]; then
@@ -49,7 +48,7 @@ conda_buildenv()
     fi
     conda config --add channels conda-forge && \
     conda config --set channel_priority strict && \
-    conda install -y "boost=1.74" "cgal=5.4" cmake ffmpeg "gdal=3.5" jupyterlab laszip "matplotlib=3.5" "mysql=8.0" "numpy=1.22" "opencv=4.5" "openmp=8.0" "pcl=1.12" "pdal=2.4" "psutil=5.9" pybind11 "qhull=2020.2" "qt=5.15.4" "scipy=1.8" sphinx_rtd_theme spyder tbb tbb-devel "xerces-c=3.2" || error_exit "conda environment ${CONDA_ENV} cannot be completed"
+    conda install -y "boost=1.74" "cgal=5.4" cmake draco ffmpeg "gdal=3.5" jupyterlab laszip "matplotlib=3.5" "mysql=8.0" "numpy=1.22" "opencv=4.5" "openmp=8.0" "pcl=1.12" "pdal=2.4" "psutil=5.9" pybind11 "qhull=2020.2" "qt=5.15.4" "scipy=1.8" sphinx_rtd_theme spyder tbb tbb-devel "xerces-c=3.2" || error_exit "conda environment ${CONDA_ENV} cannot be completed"
 }
 
 # --- CloudComPy build
@@ -78,14 +77,15 @@ cloudcompy_configure()
     -DCCCORELIB_SHARED:BOOL="1" \
     -DCCCORELIB_USE_CGAL:BOOL="1" \
     -DCCCORELIB_USE_TBB:BOOL="0" \
-    -DCMAKE_BUILD_TYPE:STRING="RelWithDebInfo" \
+    -DCMAKE_BUILD_TYPE:STRING="Release" \
     -DCMAKE_CXX_COMPILER:FILEPATH=/usr/bin/c++ \
     -DCMAKE_C_COMPILER:FILEPATH=/usr/bin/cc \
     -DCMAKE_Fortran_COMPILER:FILEPATH=/usr/bin/gfortran \
     -DCMAKE_INSTALL_PREFIX:PATH="${CLOUDCOMPY_INSTALL}" \
     -DCORK_INCLUDE_DIR:PATH="${CORK_REP}/src" \
     -DCORK_RELEASE_LIBRARY_FILE:FILEPATH="${CORK_REP}/lib/libcork.a" \
-    -DDraco_DIR:PATH="${DRACO_REP}/share/cmake" \
+    -DDRACO_INCLUDE_DIR:PATH="${CONDA_PATH}/include" \
+    -DDRACO_LIBRARY:PATH="${CONDA_PATH}/lib/libdraco.a" \
     -DEIGEN_INCLUDE_DIR:PATH="${CONDA_PATH}/include/eigen3" \
     -DEIGEN_ROOT_DIR:PATH="${CONDA_PATH}/include/eigen3" \
     -DFBX_SDK_INCLUDE_DIR:PATH="${FBXSDK_REP}/include" \